+++ rom.py
@@ -248,6 +248,7 @@ def generate_output(world: "PokemonPlatinumWorld", output_directory: str, patch:
     add_opt_byte("regional_dex_goal")
     add_opt_byte("marsh_pass")
     add_opt_byte("remote_items")
+    add_opt_byte("early_sunyshore")
 
     if len(ap_bin) % 2 == 1:
         ap_bin += b'\x00'